+++ postprocess_photos.py
@@ -440,7 +440,7 @@ def hang_around():
 
 # OK, let's go
 if __name__ == "__main__":
-    os.chdir('/home/patrick/Photos/2017-04-23')
+    
     if len(sys.argv) > 1:
         if sys.argv[1] == '--help' or sys.argv[1] == '-h':
             print_usage()